requirements/base.txt
django
pandas
psycopg2-binary
plotly
django-environ
python-dotenv
djangorestframework
django-filter
drf-yasg

